Revert "making the backend serverless for vercel"
