+++ backend/package.json
@@ -5,7 +5,7 @@
   "type": "module",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "dev": "nodemon ./api/index.js"
+    "dev": "nodemon index.js"
   },
   "repository": {
     "type": "git",
@@ -30,8 +30,7 @@
     "mongoose": "^8.18.0",
     "passport": "^0.7.0",
     "passport-local": "^1.0.0",
-    "passport-local-mongoose": "^8.0.0",
-    "serverless-http": "^4.0.0"
+    "passport-local-mongoose": "^8.0.0"
   },
   "devDependencies": {
     "nodemon": "^3.1.10"
